deployment/projects/honeysat-webpage-v2/requirements.txt
Django==4.2.13
telnetlib3==2.0.4
asgiref==3.8.1
sqlparse==0.5.0
pymongo==4.8.0
requests
python-dotenv